resources/test/tests/functional/worker-dedicated-error.html

<!DOCTYPE HTML>
<html>
<head>
<title>Dedicated Worker Error Test</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
<h1>Dedicated Web Worker Error Tests</h1>
<p>Demonstrates running <tt>testharness</tt> based tests inside a dedicated web worker, with an error.
<div id="log"></div>

<script>
test(function(t) {
        assert_true("Worker" in self, "Browser should support Workers");
    },
    "Browser supports Workers");

fetch_tests_from_worker(new Worker("worker-error.js"));
</script>
<script type="text/json" id="expected">
{
  "summarized_status": {
    "status_string": "OK",
    "message": null
  },
  "summarized_tests": [
    {
      "status_string": "PASS",
      "name": "Browser supports Workers",
      "properties": {},
      "message": null
    },
    {
      "status_string": "FAIL",
      "name": "Untitled",
      "properties": {},
      "message": "Error: This failure is expected."
    }
  ],
  "type": "complete"
}
</script>
</body>
